extract empty to a new class
