+++ frontend/packages/kubevirt-plugin/src/components/boot-order/boot-order-summary.scss
@@ -1,7 +1,7 @@
-.kubevirt-boot-order-suammary__expandable.pf-c-expandable .pf-c-expandable__content {
+.kubevirt-boot-order-summary__expandable.pf-c-expandable .pf-c-expandable__content {
   margin-top: 0;
 }
 
-.kubevirt-boot-order-suammary__empty-text {
+.kubevirt-boot-order-summary__empty-text {
   margin-bottom: 0;
 }
