+++ frontend/packages/kubevirt-plugin/src/constants/vm/constants.ts
@@ -26,9 +26,6 @@ export const VM_DETAIL_DISKS_HREF = 'disks';
 export const VM_DETAIL_NETWORKS_HREF = 'nics';
 export const VM_DETAIL_CONSOLES_HREF = 'consoles';
 
-export const DEVICE_TYPE_INTERFACE = 'interface';
-export const DEVICE_TYPE_DISK = 'disk';
-
 export const CLOUDINIT_DISK = 'cloudinitdisk';
 
 export enum DeviceType {
